+++ packages/magic-color/src/index.ts
@@ -1,7 +1,2 @@
-export * from './contrast'
-
 export * from './core'
-export * from './hash'
-export * from './random'
-export * from './theme'
 export * from '@magic-color/core'
