+++ packages/theme/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "magic-color",
+  "name": "@magic-color/theme",
   "type": "module",
   "version": "1.3.0",
-  "description": "Magic color creator.",
+  "description": "Magic color theme generator.",
   "author": "Chris <hizyyv@gmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/zyyv/magic-color#readme",
@@ -42,7 +42,6 @@
     "dev": "unbuild --stub"
   },
   "dependencies": {
-    "@magic-color/core": "workspace:*",
-    "@magic-color/theme": "workspace:*"
+    "@magic-color/core": "workspace:*"
   }
 }
